new: allow query-params to be a function
As a function, it will receive sortOrder, currentPage, and perPage
as parameters. When use as a function, Vuetable expects it to return
an object containing a key-value pair of parameters. These parameters
are used to tell the server side about the sort column, sort direction,
requested page, and page size.